requirements.txt
dbus-python
pyimgur
pypresence
pympris
python-dotenv
requests